MenuHandler/Main baseline func + Large scale refactor (#7)
* refactor(MenuHandler): Restarting Menu Handler

- Restarting menu handler
- Moved query method to SessionHandler
- changed int to double in scanner
- updated diary

* feat(MenuHandlerTest): Added skeleton tests that are failing as of now

Absolute bare-bone of tests that only expects a certain printout.
Meant to be fleshed out further as development progresses.

* feat(MenuHandler): Menu prints are mostly done, TODO: menu interaction

1 test does not pass as of yet, print all sessions.
Reason being that it does not order the list in any specific fashion.
Goal: Sort it based on relevancy/most recent

* feat(SessionHandler): getSortedSessions functional - sorts by date desc

- Base structure of sorting sessions is implemented
- will attemp to make it more flexible with enums

* feat(SortType): Enum to send into getSortedSessions

Fully functional and can be called upon/sent to get sorted method.

* refactor(Global): Refactored some parts/missed changes

Minor changes and refactoring, updated diary

* feat(Main): Main now enters MenuHandler

- Main now news up a temp user and session-collection.
- User now consists of 1 SessionHandler instance, because its logical that the collection should *belong* to someone.
- MenuHandler now takes a scanner-wrapper, and user.

* feat(Main): Main now enters MenuHandler and can run a partially working menu.

- Main can now enter MenuHandler
- MenuHandler has partially functioning methods
- Enum-UserConfig added to avoid magic strings in MenuHandler

* refactor(MenuHandlerTest): Refactored Testfile to incorporate mocking

* feat(MenuHandler): Added empty tests/failing tests to menuhandler

- Added more tests /empty methods that the branching menu will reach
- 1 is development ready.

* feat(MenuHandler): Can delete session from menu now.

deleteSessionQuery (name temp) can now be accessed and properly delete a session from the users collection

* fix(MenuHandlerTest): 3 Tests failed, fixed/adjusted to pass

old Tests from pre mocking became an issue due to us generating sessions on startup instead of per tests.

* docs(Pom): Added singular comment to pom

* build(POM): Changed version of target, removed depricated dependencies.

- commented away the arg line that conflicted with jacoco's exec build.
- arg line only removes warnings, no structual issue

* feat(MenuHandlerTest): Tests for menu branches (creating and searching session)

- also updated diary
- tests cannot run due to infinite loop because of mock "when" not fully going through, and the menu loops.

* feat(MenuHandler): sessionSearch and createSession both barebones functional

* refactor(MenuHandler): Removed redudant duplication of methods and combined it to 1

- resolveSessionView was refactored to take in a list<String> to print out and handle whatever list we gave it.
This results in us being able to use said method when we have a search query as well.

* refactor(MenuHandler): Moved 1 clear console - otherwise updated Diary

+++ src/main/java/se/iths/utility/ScannerWrapper.java
@@ -14,7 +14,7 @@ public class ScannerWrapper {
         return checkIfValidString(maxLength);
     }
 
-    public int numberInput() {
+    public double numberInput() {
         return checkIfNumber();
     }
 
@@ -59,8 +59,8 @@ public class ScannerWrapper {
         } while (true);
     }
 
-    private int checkIfNumber() {
-        int userInput;
+    private double checkIfNumber() {
+        double userInput;
         while (true) {
             String input = scanner.nextLine();
